Almost finished

+++ srcs/requirements/mariadb/tools/script.sh
@@ -1,23 +1,31 @@
 #!/bin/bash
 
 #STARTING THE MYSQL SERVICE
-service mysql start
 
-mysql_install_db --user=mysql --datadir=/var/lib/mysql
-chown mysql:mysql /run/mysqld
+if [ -d "/var/lib/mysql/$MYSQL_DDB_NAME" ]
+then
+	echo "The DDB is already installed"	
+else
+	mysql_install_db --user=mysql --datadir=/var/lib/mysql
 
-# THE '%' ALLOW REMOTE CONNECTION TO ALL HOST, AND LOCALHOST ONLY FROM THE DDB MACHINE
+	service mysql start
 
-# We delete the 'anonymous user' so you cant connect with no username
-mysql -e "DELETE FROM mysql.user WHERE User=''"
+	# THE '%' ALLOW REMOTE CONNECTION TO ALL HOST, AND LOCALHOST ONLY FROM THE DDB MACHINE
 
-mysql -e "CREATE DATABASE IF NOT EXISTS $MYSQL_DDB_NAME;"
+	# We delete the 'anonymous user' so you cant connect with no username
+	mysql -e "DELETE FROM mysql.user WHERE User=''"
 
-mysql -e "CREATE USER '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';"
+	mysql -e "CREATE DATABASE IF NOT EXISTS $MYSQL_DDB_NAME;"
 
-mysql -e "GRANT ALL PRIVILEGES ON $MYSQL_DDB_NAME.* TO '$MYSQL_USER'@'%';"
+	mysql -e "CREATE USER '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD';"
 
-# We apply our changement before changing root password, otherwise the script wont be able to exec the remaining commands
-mysql -e "FLUSH PRIVILEGES;"
+	mysql -e "GRANT ALL PRIVILEGES ON $MYSQL_DDB_NAME.* TO '$MYSQL_USER'@'%';"
 
-mysql -e "ALTER USER 'root'@'localhost' IDENTIFIED BY '$MYSQL_ROOT_PASSWORD';"
+	# We apply our changement before changing root password, otherwise the script wont be able to exec the remaining commands
+	mysql -e "FLUSH PRIVILEGES;"
+
+	# Then we change root password
+	mysqladmin -u root password $MYSQL_ROOT_PASSWORD
+fi
+
+mysqld_safe
